Implement presentational listing navigation
Navigation is shown below the header. The component supports a back button for
use when we're displaying documents of a course. The text and back button
arrange themselves BEAUTIFULLY even when on a super narrow (360px) phone screen,
moving the text below the back button like seen in some Material Design
examples.