Fix unicode filenames in Content-Disposition
Express doesn't apparently do anything with the strings passed to header values,
so we need to manually provide a non-utf-8 filename in Content-Disposition. I've
used the `content-disposition` package which handles the proper header structure
including a non-utf-8 fallback option. Unfortunately the package replaces
non-ASCII characters with question marks, which aren't valid filenames on
Windows, so I've used the `transliteration` package to transliterate the
probable åäö-characters etc. to ASCII.