Fix db seeder script not updating ID sequences
Previously the db seeder script would generate IDs for courses and exams itself,
which meant that the primary key sequences would not be updated in the database.

Now, the script inserts the items without IDs, allowing the DB to pick the IDs
itself. This also means that the script doesn't need to be run on an empty
database anymore, but that it can be used to migrate data from the old tarpisto
to the new archive even if the new archive has been in production.

Additionally, the script is now more robust and if it fails, it reports what
kind of resources have been left dirty (where were rows inserted, files copied).