Add dev route for AWS_S3_DEV_PREFIX
Instead of applying the dev prefix manually to the object key names,
apply the dev prefix to the exams' file_path in the database and in the
multer S3 object key generator.

Adds new dev route under /dev which allows you to inspect the db for
existing dev prefixes in the file_path and apply a new one or remove the
prefix completely. Only enabled when NODE_ENV === 'development'.