Implement exam file rename functionality
Now the exam is renamed in both the DB and S3. The S3 metadata
(filename in content-disposition) needs to be changed by copying the
object. The old filename is cached by CloudFront for some time, but if
it becomes a problem, we can just generate a new UUID for the object so
CloudFront is forced to fetch the object from S3.