Ignore view jsx in nodemon
Express view caching is not enabled when NODE_ENV===development, so
ignore view JSX files in nodemon so that the server doesn't need to
restart every time a view file is modified. This way we can just refresh
the page to see the changes, instead of waiting for the server to
restart.